Fix ResourcePool.tryGetFeatureType()
Catch and log all exceptions and properly flush the feature type
held by the data store associated with a FeatureTypeInfo to be safe
in case the underlying schema has changed.

So far this method only works with ContentDataStore instances
(i.e. all JDBC ones and others, but not all). This is to avoid calling
DataStore.dispose() as other threads may be using it and has proved to
result in unpredictable errors. Instead, we're calling the datastore
feature type's ContentState.flush() method which forces re-loading
the native type when next used.